new GlobalItemList script where all items are stored. New ItemLoader script WIP :D :3
