+++ Assets/Scripts/GlobalItemList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cee39ea3c99cf704aa03443b66f41a64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
